# swagger.yaml
openapi: 3.0.3
info:
  title: Restaurant Food Ordering System API
  description: REST API for the Restaurant Food Ordering System
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MenuCategoryType:
      type: string
      enum: [VEG, NON_VEG, DRINKS, SPECIALS, DESSERTS, APPETIZERS, SIDES]
    ItemAvailability:
      type: string
      enum: [AVAILABLE, OUT_OF_STOCK, HIDDEN]
    OrderStatus:
      type: string
      enum: [PENDING, CONFIRMED, PREPARING, READY, COMPLETED, CANCELLED]
...
    SalaryStatus:
      type: string
      enum: [PENDING, PAID, ON_HOLD]
...
    MenuCategoryResponse:
      type: object
      properties:
        code:          { type: string }
        restaurantCode: { type: string }
        name:          { type: string }
        categoryType:  { $ref: '#/components/schemas/MenuCategoryType' }
        fileCode:      { type: string, nullable: true }
        sortOrder:     { type: integer }
        isActive:      { type: boolean }
        createAt:      { type: string, format: date-time }
    MenuItemResponse:
      type: object
      properties:
        code:            { type: string }
        restaurantCode:  { type: string }
        categoryCode:    { type: string, nullable: true }
        name:            { type: string }
        description:     { type: string, nullable: true }
        price:           { type: number }
        discountPercent: { type: number }
        fileCode:        { type: string, nullable: true }
        availability:    { $ref: '#/components/schemas/ItemAvailability' }
        isFeatured:      { type: boolean }
        isVeg:           { type: boolean }
        prepTimeMinutes: { type: integer }
        sortOrder:       { type: integer }
        createAt:        { type: string, format: date-time }
...
    OrderItemResponse:
      type: object
      properties:
        code:           { type: string }
        orderCode:      { type: string }
        menuItemCode:   { type: string }
        quantity:       { type: integer }
        unitPrice:      { type: number }
        discountAmount: { type: number }
        totalPrice:     { type: number }
        spiceLevel:     { type: string, nullable: true }
        notes:          { type: string, nullable: true }
        status:         { $ref: '#/components/schemas/OrderStatus' }
        createAt:       { type: string, format: date-time }
...
    EmployeeResponse:
      type: object
      properties:
        code:           { type: string }
        restaurantCode: { type: string }
        fileCode:       { type: string, nullable: true }
        fullName:       { type: string }
        phone:          { type: string, nullable: true }
        joinDate:       { type: string, format: date }
        baseSalary:     { type: number }
        bankAccount:    { type: string, nullable: true }
        bankName:       { type: string, nullable: true }
        isActive:       { type: boolean }
        createAt:       { type: string, format: date-time }
...
    PayrollResponse:
      type: object
      properties:
        code:           { type: string }
        restaurantCode: { type: string }
        employeeCode:   { type: string }
        payPeriodStart: { type: string, format: date }
        payPeriodEnd:   { type: string, format: date }
        overtimePay:    { type: number }
        bonus:          { type: number }
        deductions:     { type: number }
        netSalary:      { type: number }
        status:         { $ref: '#/components/schemas/SalaryStatus' }
        paidAt:         { type: string, format: date-time, nullable: true }
        createAt:       { type: string, format: date-time }
...
    PageMeta:
      type: object
      properties:
        totalElements: { type: integer }
        totalPages:    { type: integer }
        size:          { type: integer }
        number:        { type: integer }
  parameters:
    PageParam:
      name: page
      in: query
      schema: { type: integer, default: 0 }
    SizeParam:
      name: size
      in: query
      schema: { type: integer, default: 20 }

# ═══════════════════════════════════════════════════════════════
# PATHS
# ═══════════════════════════════════════════════════════════════
paths:
...
  /menu-categories:
    post:
      tags: [Menu Categories]
      summary: Create menu category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [restaurantCode, name]
              properties:
                restaurantCode: { type: string }
                name:           { type: string }
                categoryType:   { $ref: '#/components/schemas/MenuCategoryType' }
                sortOrder:      { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MenuCategoryResponse' }
  /menu-categories/search:
    get:
      tags: [Menu Categories]
      summary: Search menu categories (public)
      security: []
      parameters:
        - { name: restaurantCode, in: query, schema: { type: string } }
        - { name: categoryType, in: query, schema: { $ref: '#/components/schemas/MenuCategoryType' } }
        - { $ref: '#/components/parameters/PageParam' }
        - { $ref: '#/components/parameters/SizeParam' }
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/PageMeta' }
                  - type: object
                    properties:
                      content:
                        type: array
                        items: { $ref: '#/components/schemas/MenuCategoryResponse' }
  /menu-categories/{code}:
    get:
      tags: [Menu Categories]
      summary: Get menu category
      security: []
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MenuCategoryResponse' }
    patch:
      tags: [Menu Categories]
      summary: Update menu category
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:         { type: string }
                categoryType: { $ref: '#/components/schemas/MenuCategoryType' }
                sortOrder:    { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MenuCategoryResponse' }
    delete:
      tags: [Menu Categories]
      summary: Delete menu category
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Deleted }

  # ── MENU ITEMS ───────────────────────────────────────────────
  /menu-items:
    post:
      tags: [Menu Items]
      summary: Create menu item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [restaurantCode, name, price]
              properties:
                restaurantCode:  { type: string }
                categoryCode:    { type: string }
                name:            { type: string }
                description:     { type: string }
                price:           { type: number }
                discountPercent: { type: number, default: 0 }
                fileCode:        { type: string }
                isVeg:           { type: boolean }
                prepTimeMinutes: { type: integer }
                sortOrder:       { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MenuItemResponse' }
  /menu-items/search:
    get:
      tags: [Menu Items]
      summary: Search menu items (public)
      security: []
      parameters:
        - { name: restaurantCode, in: query, schema: { type: string } }
        - { name: categoryCode, in: query, schema: { type: string } }
        - { name: availability, in: query, schema: { $ref: '#/components/schemas/ItemAvailability' } }
        - { name: isFeatured, in: query, schema: { type: boolean } }
        - { name: isVeg, in: query, schema: { type: boolean } }
        - { $ref: '#/components/parameters/PageParam' }
        - { $ref: '#/components/parameters/SizeParam' }
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/PageMeta' }
                  - type: object
                    properties:
                      content:
                        type: array
                        items: { $ref: '#/components/schemas/MenuItemResponse' }
  /menu-items/{code}:
    get:
      tags: [Menu Items]
      summary: Get menu item
      security: []
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MenuItemResponse' }
    patch:
      tags: [Menu Items]
      summary: Update menu item
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                categoryCode:    { type: string }
                name:            { type: string }
                description:     { type: string }
                price:           { type: number }
                discountPercent: { type: number }
                fileCode:        { type: string }
                availability:    { $ref: '#/components/schemas/ItemAvailability' }
                isFeatured:      { type: boolean }
                isVeg:           { type: boolean }
                prepTimeMinutes: { type: integer }
                sortOrder:       { type: integer }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MenuItemResponse' }
    delete:
      tags: [Menu Items]
      summary: Delete menu item
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Deleted }

  # ── ORDERS ───────────────────────────────────────────────────
...
  /orders/{orderCode}/items/{code}:
    patch:
      tags: [Order Items]
      summary: Update order item
      parameters:
        - { name: orderCode, in: path, required: true, schema: { type: string } }
        - { name: code, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:   { type: integer }
                spiceLevel: { type: string }
                notes:      { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderItemResponse' }
    delete:
      tags: [Order Items]
      summary: Remove order item
      parameters:
        - { name: orderCode, in: path, required: true, schema: { type: string } }
        - { name: code, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Removed }

  # ── PAYMENTS ─────────────────────────────────────────────────
...
  /employees:
    post:
      tags: [Employees]
      summary: Create employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [restaurantCode, fullName, joinDate, baseSalary]
              properties:
                restaurantCode: { type: string }
                fullName:       { type: string }
                phone:          { type: string }
                joinDate:       { type: string, format: date }
                baseSalary:     { type: number }
                bankAccount:    { type: string }
                bankName:       { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeeResponse' }
  /employees/search:
    get:
      tags: [Employees]
      summary: Search employees
      parameters:
        - { name: restaurantCode, in: query, schema: { type: string } }
        - { name: fullName, in: query, schema: { type: string } }
        - { name: isActive, in: query, schema: { type: boolean } }
        - { $ref: '#/components/parameters/PageParam' }
        - { $ref: '#/components/parameters/SizeParam' }
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/PageMeta' }
                  - type: object
                    properties:
                      content:
                        type: array
                        items: { $ref: '#/components/schemas/EmployeeResponse' }
  /employees/{code}:
    get:
      tags: [Employees]
      summary: Get employee
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeeResponse' }
    patch:
      tags: [Employees]
      summary: Update employee
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                fullName:    { type: string }
                phone:       { type: string }
                baseSalary:  { type: number }
                bankAccount: { type: string }
                bankName:    { type: string }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EmployeeResponse' }
    delete:
      tags: [Employees]
      summary: Delete employee
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      responses:
        '204': { description: Deleted }

  # ── ATTENDANCE ───────────────────────────────────────────────
...
  /payroll:
    post:
      tags: [Payroll]
      summary: Create payroll record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [restaurantCode, employeeCode, payPeriodStart, payPeriodEnd, netSalary]
              properties:
                restaurantCode: { type: string }
                employeeCode:   { type: string }
                payPeriodStart: { type: string, format: date }
                payPeriodEnd:   { type: string, format: date }
                overtimePay:    { type: number, default: 0 }
                bonus:          { type: number, default: 0 }
                deductions:     { type: number, default: 0 }
                netSalary:      { type: number }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayrollResponse' }
  /payroll/search:
    get:
      tags: [Payroll]
      summary: Search payroll records
      parameters:
        - { name: restaurantCode, in: query, schema: { type: string } }
        - { name: employeeCode, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/SalaryStatus' } }
        - { $ref: '#/components/parameters/PageParam' }
        - { $ref: '#/components/parameters/SizeParam' }
      responses:
        '200':
          content:
            application/json:
              schema:
                allOf:
                  - { $ref: '#/components/schemas/PageMeta' }
                  - type: object
                    properties:
                      content:
                        type: array
                        items: { $ref: '#/components/schemas/PayrollResponse' }
  /payroll/{code}:
    get:
      tags: [Payroll]
      summary: Get payroll record
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayrollResponse' }
  /payroll/{code}/status:
    patch:
      tags: [Payroll]
      summary: Update payroll status
      parameters:
        - { name: code, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/SalaryStatus' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PayrollResponse' }

  # ── USERS ────────────────────────────────────────────────────
...